+++ src/libslic3r/Print.hpp
@@ -178,7 +178,7 @@ private:
     void generate_support_material();
 
     void _slice(const std::vector<coordf_t> &layer_height_profile);
-    void _offset_holes(float hole_delta, LayerRegion *layerm);
+    void _offset_holes(double hole_delta, LayerRegion *layerm);
     void _smooth_curves(LayerRegion *layerm);
     std::string _fix_slicing_errors();
     void _simplify_slices(double distance);
